Improve, and finalize GitHub workflows
Because:
- Helm GitHub Actions required more discovery work
- A workflow for managing stale issues and PRs was needed
- Pinning GitHub Actions to their major version is safe enough

The documentation of the new Helm GitHub Actions leave some healthy
room for improvement and clarification, which necessitate doing some
experimentation, and fine tuning.

A new custom, but simple workflow for managing stale issues and PRs
is added.

Given that the majority of used GitHub Actions are verified, the
resolution of version pinning is reduced to their major version.

Changes to be committed:
	deleted:    .github/stale.yml
	modified:   .github/workflows/helm-release.yml
	modified:   .github/workflows/oci-image-build.yml

Untracked files:
	.github/workflows/stale.yaml